# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MTS Transport Service API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Route {
  id: number;
  title: string;
  fromLocation: string;
  toLocation: string;
  covered: string;
  timing: string;
  vehicle: string;
  category: string;
  createdAt: string;
}

export interface CreateRouteBody {
  title: string;
  fromLocation: string;
  toLocation: string;
  covered?: string;
  timing?: string;
  vehicle: string;
  category: string;
}

export interface UpdateRouteBody {
  title?: string;
  fromLocation?: string;
  toLocation?: string;
  covered?: string;
  timing?: string;
  vehicle?: string;
  category?: string;
}

export interface Message {
  id: number;
  name: string;
  phone: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface CreateMessageBody {
  name: string;
  phone: string;
  message: string;
}

export interface UpdateMessageBody {
  status: string;
}

export interface ContactRequest {
  id: number;
  name: string;
  phone: string;
  pickupLocation: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  phone: string;
  pickupLocation?: string;
  message?: string;
}

export interface UpdateContactBody {
  status: string;
}