# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MTS Transport Service API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all routes
 */
export const GetRoutesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  fromLocation: zod.string(),
  toLocation: zod.string(),
  covered: zod.string(),
  timing: zod.string(),
  vehicle: zod.string(),
  category: zod.string(),
  createdAt: zod.string(),
});
export const GetRoutesResponse = zod.array(GetRoutesResponseItem);

/**
 * @summary Create a route
 */
export const CreateRouteBody = zod.object({
  title: zod.string(),
  fromLocation: zod.string(),
  toLocation: zod.string(),
  covered: zod.string().optional(),
  timing: zod.string().optional(),
  vehicle: zod.string(),
  category: zod.string(),
});

/**
 * @summary Update a route
 */
export const UpdateRouteParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateRouteBody = zod.object({
  title: zod.string().optional(),
  fromLocation: zod.string().optional(),
  toLocation: zod.string().optional(),
  covered: zod.string().optional(),
  timing: zod.string().optional(),
  vehicle: zod.string().optional(),
  category: zod.string().optional(),
});

export const UpdateRouteResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  fromLocation: zod.string(),
  toLocation: zod.string(),
  covered: zod.string(),
  timing: zod.string(),
  vehicle: zod.string(),
  category: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a route
 */
export const DeleteRouteParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all messages
 */
export const GetMessagesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  message: zod.string(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const GetMessagesResponse = zod.array(GetMessagesResponseItem);

/**
 * @summary Create a message
 */
export const CreateMessageBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  message: zod.string(),
});

/**
 * @summary Update message status
 */
export const UpdateMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMessageBody = zod.object({
  status: zod.string(),
});

export const UpdateMessageResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  message: zod.string(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a message
 */
export const DeleteMessageParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all contact requests
 */
export const GetContactsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  pickupLocation: zod.string(),
  message: zod.string(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const GetContactsResponse = zod.array(GetContactsResponseItem);

/**
 * @summary Submit a contact request
 */
export const CreateContactBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  pickupLocation: zod.string().optional(),
  message: zod.string().optional(),
});

/**
 * @summary Update contact request status
 */
export const UpdateContactParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateContactBody = zod.object({
  status: zod.string(),
});

export const UpdateContactResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  pickupLocation: zod.string(),
  message: zod.string(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a contact request
 */
export const DeleteContactParams = zod.object({
  id: zod.coerce.number(),
});